no need in separate package for reducer
